変更 - ネストしているオブジェクトを、必要な要素だけで平坦化するのをモジュールとして分離
