+++ app/_utile/prisma/index.ts
@@ -1,67 +1,11 @@
-import { HikasenVtuber, Tag } from '@/(types)';
+import { HikasenVtuber, Tags, Tag } from '@/(types)';
 import { PrismaClient, Prisma } from '@prisma/client';
 
-const prisma = new PrismaClient({
+import { convertTaggingToTags } from '@/_utile/convert';
+
+export const prisma = new PrismaClient({
   log: ['query', 'error', 'info', 'warn'],
 });
-
-/**
- * プレイスタイルや配信スタイルなどの属性タグを、カテゴリー毎に分類しオブジェクトにする
- * @param channels PrismaChannel
- * @returns
- */
-const convertTags = (channels): HikasenVtuber[] => {
-  return channels.map((channel) => {
-    //DBのタグ情報は、1つのテーブルに入っているので、種類毎に使いやすいように分割する
-    const contents: Tag[] = [];
-    const party: Tag[] = [];
-    const timezone: Tag[] = [];
-    channel.tags.forEach((tag) => {
-      //ソートのために必要なID、表示に必要な名前、コードのみにオブジェクトを限定する
-      const convert = {
-        id: tag.tags.id,
-        name: tag.tags.name,
-        code: tag.tags.code,
-      };
-
-      //見やすいようにSwitch文で切り分ける
-      switch (tag.tags.type) {
-        case 'content':
-          contents.push(convert);
-          break;
-        case 'play':
-          party.push(convert);
-          break;
-        case 'timezone':
-          timezone.push(convert);
-          break;
-        default:
-          break;
-      }
-    });
-
-    //サーバー登録順は必ずしもID順に並んでいないので、ID順に並べ直させる
-    contents.sort((a, b) => {
-      return a.id - b.id;
-    });
-    party.sort((a, b) => {
-      return a.id - b.id;
-    });
-    timezone.sort((a, b) => {
-      return a.id - b.id;
-    });
-
-    //
-    return {
-      ...channel,
-      tags: {
-        content: contents,
-        party: party,
-        timezone: timezone,
-      },
-    };
-  });
-};
 
 /**
  * 配信者一覧を降順で無条件取得をする
@@ -89,7 +33,7 @@ export const getChannelOffset = async (
     },
   });
 
-  const result = convertTags(channels);
+  const result = convertTaggingToTags(channels);
 
   return result;
 };
@@ -132,7 +76,7 @@ export const getChannelWhereOffset = async (
     },
   });
 
-  const result = convertTags(channels);
+  const result = convertTaggingToTags(channels);
 
   return result;
 };
